Crate substrate_header_metadata

Implements tree backend, cached header metadata and algorithms to compute routes efficiently over the tree of headers.

Structs

CachedHeaderMetadata
Cached header metadata. Used to efficiently traverse the tree.

HashAndNumber
Hash and number of a block.

HeaderMetadataCache
Caches header metadata in an in-memory LRU cache.

TreeRoute
A tree-route from one block to another in the chain.

Traits

HeaderMetadata
Handles header metadata: hash, number, parent hash, etc.

Functions

lowest_common_ancestor
Get lowest common ancestor between two blocks in the tree.

tree_route
Compute a tree-route between two blocks. See tree-route docs for more details.